Correct services for property and feedback

+++ backEnd/src/main/java/tn/esprit/backend/repository/PropertyRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.backend.entity.Property;
 
-public interface PropertyRepository extends JpaRepository<Property,Integer> {
+public interface PropertyRepository extends JpaRepository<Property,Long> {
 }
